FlareSolverr process is now terminated more gracefully Previously the Kill method was used to terminate the FlareSolverr process. However, this often left browser instances behind. The CloseMainWindow method is now used instead, allowing it to shutdown gracefully.

+++ src/Gsemac.Net.Cloudflare/FlareSolverr/FlareSolverrService.cs
@@ -209,7 +209,11 @@ namespace Gsemac.Net.Cloudflare.FlareSolverr {
 
                 logger.Info("Stopping FlareSolverr process");
 
-                flareSolverrProcess.Kill();
+                // Attempt to stop the process gracefully so that it can close running browser instances.
+
+                if (!flareSolverrProcess.CloseMainWindow())
+                    flareSolverrProcess.Kill();
+
                 flareSolverrProcess.Dispose();
 
                 flareSolverrProcess = null;
